Refactor utility functions (#55)
* move operations and transform util files into a dedicated utils folder

* move testing util files to a new test_utils folder

* move assert_melt function to test_utils folder

* update util imports in all files

* update util import statements in tests

* split util files into better named files

* split util files into new test files

* fix assertion tolerance

* fix np builtin type deprecation in 1.2

* revert back to deprecated check_less_precise

* add util init file

* Update cytominer_eval/tests/test_operations/test_mp_value.py


* Update cytominer_eval/tests/test_utils/test_availability_utils.py


* Update cytominer_eval/tests/test_utils/test_grit_utils.py


* Apply suggestions from code review

Fixing spaces after function import, docstring consistency, and a missing numpy builtin update

+++ cytominer_eval/evaluate.py
@@ -8,7 +8,7 @@ import pandas as pd
 from typing import List, Union
 
 from cytominer_eval.transform import metric_melt
-from cytominer_eval.transform.util import check_replicate_groups
+from cytominer_eval.utils.transform_utils import check_replicate_groups
 from cytominer_eval.operations import (
     replicate_reproducibility,
     precision_recall,
@@ -25,7 +25,7 @@ def evaluate(
     replicate_groups: Union[List[str], dict],
     operation: str = "replicate_reproducibility",
     similarity_metric: str = "pearson",
-    replicate_reproducibility_quantile: np.float = 0.95,
+    replicate_reproducibility_quantile: float = 0.95,
     replicate_reproducibility_return_median_cor: bool = False,
     precision_recall_k: Union[int, List[int]] = 10,
     grit_control_perts: List[str] = ["None"],
